requirements.txt
# ─── Core Data & ML ───────────────────────────────────────────────────────────
streamlit>=1.32
pandas>=2.0
numpy>=1.26
scikit-learn>=1.4
scipy>=1.12
joblib>=1.3

# ─── Visualization ────────────────────────────────────────────────────────────
plotly>=5.20
matplotlib>=3.8
seaborn>=0.13

# ─── Explainability ───────────────────────────────────────────────────────────
shap>=0.44

# ─── Experiment Tracking ──────────────────────────────────────────────────────
mlflow>=2.11

# ─── Backend / API ────────────────────────────────────────────────────────────
fastapi>=0.110
uvicorn[standard]>=0.29
httpx>=0.27
pydantic>=2.6
python-dotenv>=1.0

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0

# ─── LLM Integration ──────────────────────────────────────────────────────────
google-generativeai>=0.5
requests>=2.31

# ─── Reporting ────────────────────────────────────────────────────────────────
openpyxl>=3.1
reportlab>=4.1
Pillow>=10.0

# ─── Testing / Lint ───────────────────────────────────────────────────────────
pytest>=8.0
flake8>=7.0